+++ mystdlib/src/mystrlen.c
@@ -1,9 +1,9 @@
 #include <libio.h>
 #include <mystring.h>
 
-int mystrlen(const char *s) {
+size_t mystrlen(const char *s) {
 
-    int i = 0;
+    size_t i = 0;
     if (NULL != s) {
         while ('\0' != s[i]) {
             ++i;
